Return a value from a loop
This is a contrived example.
The loop from the previous example is now in a
function. The function is declared to return an
integer. The previous break has been replaced
with a return. It returns the same value as
before. The return value is placed in an
immutable variable and printed.