Initial commit
The program consists of one main loop. So the main
function reflects this with a body that loops on
the "ask" function. This function returns a
boolean true to continue and false to finish.
The ! for negation of the function return must
be separated from the function name or it would
look like a macro.

In the ask function, the user input is placed into
a mutable string. Subsequent incorrect use of this
causes it to be marked at the declaration with a
warning or error message. Look further down for
the reason.

The mutable string is CLONED using a reference
so that the rest of the program cannot change it.
The input is reported. If it is "q" the loop
quits.

The input is tokenized. If the tokenization fails
the loop is continued.

The tokenizer takes a reference to the input
String and returns references to string slices.
These refer to the input string.