Add tokenizer
The tokenizer splits the input string on
whitespace. The result is placed in a vector.
Each vector item is an Option, so they are
extracted using expect. This of course succeeds
because we know the size of the vector after
checking it.

If the vector is not the expected length, the
returned tuple uses false to indicate that
the input was not understood. We allow either
10 C or F. The temperature is also parsed as f64.